Convert `twap_length_seconds` to a `u32`
This means we can't overflow when casting to i64

+++ programs/price_based_performance_package/src/instructions/initialize_performance_package.rs
@@ -9,7 +9,7 @@ pub struct InitializePerformancePackageParams {
     pub tranches: Vec<Tranche>,
     pub min_unlock_timestamp: i64,
     pub oracle_config: OracleConfig,
-    pub twap_length_seconds: u64,
+    pub twap_length_seconds: u32,
     pub grantee: Pubkey,
     pub performance_package_authority: Pubkey,
 }
